allow VCR options to be supplied via metadata
Thanks for minitest-vcr!

I wanted to be able to quickly flip into record mode, so added the
ability to supply a hash of options.

I also made two changes:

1. removed database cleaner. I couldn't see any reason this
functionality was related to VCR, and seemed to be likely to suprise
people
1. removed extension to string. Seemed like a very big/surprising hammer
to use for 3 method calls in one place